+++ src/allmydata/test/no_network.py
@@ -30,7 +30,7 @@ from allmydata.util import fileutil, idlib, hashutil
 from allmydata.util.hashutil import sha1
 from allmydata.test.common_web import HTTPClientGETFactory
 from allmydata.interfaces import IStorageBroker, IServer
-from allmydata.test.common import TEST_RSA_KEY_SIZE
+from .common import TEST_RSA_KEY_SIZE
 
 
 class IntentionalError(Exception):
